Add timeline utilities package
Extract Studio timeline strip helpers into @remotion/timeline-utils so filmstrip rendering and loop calculations can be shared outside Studio.
