+++ packages/studio/src/audio-waveform-worker.ts
@@ -1,5 +1,9 @@
 /// <reference lib="webworker" />
 
+import {
+	getLoopDisplayWidth,
+	shouldTileLoopDisplay,
+} from '@remotion/timeline-utils';
 import type {
 	AudioWaveformWorkerIncomingMessage,
 	AudioWaveformWorkerOutgoingMessage,
@@ -7,10 +11,6 @@ import type {
 } from './components/audio-waveform-worker-types';
 import {drawBars} from './components/draw-peaks';
 import {loadWaveformPeaks} from './components/load-waveform-peaks';
-import {
-	getLoopDisplayWidth,
-	shouldTileLoopDisplay,
-} from './components/looped-media-timeline';
 import {sliceWaveformPeaks} from './components/slice-waveform-peaks';
 
 declare const self: DedicatedWorkerGlobalScope;
