+++ packages/studio/src/components/Timeline/TimelineVideoInfo.tsx
@@ -1,27 +1,25 @@
-import React, {useEffect, useMemo, useRef, useState} from 'react';
-import type {LoopDisplay} from 'remotion';
-import {useVideoConfig} from 'remotion';
-import {extractFrames} from '../../helpers/extract-frames';
-import type {FrameDatabaseKey} from '../../helpers/frame-database';
 import {
 	addFrameToCache,
 	aspectRatioCache,
-	frameDatabase,
+	ensureSlots,
+	extractFrames,
+	fillFrameWhereItFits,
+	fillWithCachedFrames,
+	getLoopDisplayWidth,
 	getAspectRatioFromCache,
-	getFrameDatabaseKeyPrefix,
-	getTimestampFromFrameDatabaseKey,
 	makeFrameDatabaseKey,
-} from '../../helpers/frame-database';
-import {resizeVideoFrame} from '../../helpers/resize-video-frame';
+	resizeVideoFrame,
+	shouldTileLoopDisplay,
+	WEBCODECS_TIMESCALE,
+} from '@remotion/timeline-utils';
+import React, {useEffect, useMemo, useRef, useState} from 'react';
+import type {LoopDisplay} from 'remotion';
+import {useVideoConfig} from 'remotion';
 import {
 	TIMELINE_LAYER_HEIGHT_AUDIO,
 	TIMELINE_LAYER_HEIGHT_IMAGE,
 } from '../../helpers/timeline-layout';
 import {AudioWaveform} from '../AudioWaveform';
-import {
-	getLoopDisplayWidth,
-	shouldTileLoopDisplay,
-} from '../looped-media-timeline';
 
 const FILMSTRIP_HEIGHT = TIMELINE_LAYER_HEIGHT_IMAGE - 2;
 
@@ -40,245 +38,6 @@ const filmstripContainerStyle: React.CSSProperties = {
 	borderTopLeftRadius: 2,
 	fontSize: 10,
 	fontFamily: 'Arial, Helvetica',
-};
-
-const WEBCODECS_TIMESCALE = 1_000_000;
-const MAX_TIME_DEVIATION = WEBCODECS_TIMESCALE * 0.05;
-
-const getDurationOfOneFrame = ({
-	visualizationWidth,
-	aspectRatio,
-	segmentDuration,
-}: {
-	visualizationWidth: number;
-	aspectRatio: number;
-	segmentDuration: number;
-}) => {
-	const framesFitInWidthUnrounded =
-		visualizationWidth / (FILMSTRIP_HEIGHT * aspectRatio);
-	return (segmentDuration / framesFitInWidthUnrounded) * WEBCODECS_TIMESCALE;
-};
-
-const fixRounding = (value: number) => {
-	if (value % 1 >= 0.49999999) {
-		return Math.ceil(value);
-	}
-
-	return Math.floor(value);
-};
-
-const calculateTimestampSlots = ({
-	visualizationWidth,
-	fromSeconds,
-	segmentDuration,
-	aspectRatio,
-}: {
-	visualizationWidth: number;
-	fromSeconds: number;
-	segmentDuration: number;
-	aspectRatio: number;
-}) => {
-	const framesFitInWidthUnrounded =
-		visualizationWidth / (FILMSTRIP_HEIGHT * aspectRatio);
-	const framesFitInWidth = Math.ceil(framesFitInWidthUnrounded);
-	const durationOfOneFrame = getDurationOfOneFrame({
-		visualizationWidth,
-		aspectRatio,
-		segmentDuration,
-	});
-
-	const timestampTargets: number[] = [];
-	for (let i = 0; i < framesFitInWidth + 1; i++) {
-		const target =
-			fromSeconds * WEBCODECS_TIMESCALE + durationOfOneFrame * (i + 0.5);
-		const snappedToDuration =
-			(Math.round(fixRounding(target / durationOfOneFrame)) - 1) *
-			durationOfOneFrame;
-
-		timestampTargets.push(snappedToDuration);
-	}
-
-	return timestampTargets;
-};
-
-const ensureSlots = ({
-	filledSlots,
-	naturalWidth,
-	fromSeconds,
-	toSeconds,
-	aspectRatio,
-}: {
-	filledSlots: Map<number, number | undefined>;
-	naturalWidth: number;
-	fromSeconds: number;
-	toSeconds: number;
-	aspectRatio: number;
-}) => {
-	const segmentDuration = toSeconds - fromSeconds;
-
-	const timestampTargets = calculateTimestampSlots({
-		visualizationWidth: naturalWidth,
-		fromSeconds,
-		segmentDuration,
-		aspectRatio,
-	});
-
-	for (const timestamp of timestampTargets) {
-		if (!filledSlots.has(timestamp)) {
-			filledSlots.set(timestamp, undefined);
-		}
-	}
-};
-
-const drawSlot = ({
-	frame,
-	ctx,
-	filledSlots,
-	visualizationWidth,
-	timestamp,
-	segmentDuration,
-	fromSeconds,
-}: {
-	frame: VideoFrame;
-	ctx: CanvasRenderingContext2D;
-	filledSlots: Map<number, number | undefined>;
-	visualizationWidth: number;
-	timestamp: number;
-	segmentDuration: number;
-	fromSeconds: number;
-}) => {
-	const durationOfOneFrame = getDurationOfOneFrame({
-		visualizationWidth,
-		aspectRatio: frame.displayWidth / frame.displayHeight,
-		segmentDuration,
-	});
-
-	const relativeTimestamp = timestamp - fromSeconds * WEBCODECS_TIMESCALE;
-	const frameIndex = relativeTimestamp / durationOfOneFrame;
-	const thumbnailWidth = frame.displayWidth / window.devicePixelRatio;
-	const left = Math.floor(frameIndex * thumbnailWidth);
-	const right = Math.ceil((frameIndex + 1) * thumbnailWidth);
-
-	ctx.drawImage(
-		frame,
-		left,
-		0,
-		right - left,
-		frame.displayHeight / window.devicePixelRatio,
-	);
-	filledSlots.set(timestamp, frame.timestamp);
-};
-
-const fillWithCachedFrames = ({
-	ctx,
-	naturalWidth,
-	filledSlots,
-	src,
-	segmentDuration,
-	fromSeconds,
-}: {
-	ctx: CanvasRenderingContext2D;
-	naturalWidth: number;
-	filledSlots: Map<number, number | undefined>;
-	src: string;
-	segmentDuration: number;
-	fromSeconds: number;
-}) => {
-	const prefix = getFrameDatabaseKeyPrefix(src);
-	const keys = Array.from(frameDatabase.keys()).filter((k) =>
-		k.startsWith(prefix),
-	);
-	const targets = Array.from(filledSlots.keys());
-
-	for (const timestamp of targets) {
-		let bestKey: FrameDatabaseKey | undefined;
-		let bestDistance = Infinity;
-		for (const key of keys) {
-			const distance = Math.abs(
-				getTimestampFromFrameDatabaseKey(key) - timestamp,
-			);
-			if (distance < bestDistance) {
-				bestDistance = distance;
-				bestKey = key;
-			}
-		}
-
-		if (!bestKey) {
-			continue;
-		}
-
-		const frame = frameDatabase.get(bestKey);
-		if (!frame) {
-			continue;
-		}
-
-		const alreadyFilled = filledSlots.get(timestamp);
-
-		// Don't fill if a closer frame was already drawn
-		if (
-			alreadyFilled &&
-			Math.abs(alreadyFilled - timestamp) <=
-				Math.abs(frame.frame.timestamp - timestamp)
-		) {
-			continue;
-		}
-
-		frame.lastUsed = Date.now();
-
-		drawSlot({
-			ctx,
-			frame: frame.frame,
-			filledSlots,
-			visualizationWidth: naturalWidth,
-			timestamp,
-			segmentDuration,
-			fromSeconds,
-		});
-	}
-};
-
-const fillFrameWhereItFits = ({
-	frame,
-	filledSlots,
-	ctx,
-	visualizationWidth,
-	segmentDuration,
-	fromSeconds,
-}: {
-	frame: VideoFrame;
-	filledSlots: Map<number, number | undefined>;
-	ctx: CanvasRenderingContext2D;
-	visualizationWidth: number;
-	segmentDuration: number;
-	fromSeconds: number;
-}) => {
-	const slots = Array.from(filledSlots.keys());
-
-	for (let i = 0; i < slots.length; i++) {
-		const slot = slots[i];
-		if (Math.abs(slot - frame.timestamp) > MAX_TIME_DEVIATION) {
-			continue;
-		}
-
-		const filled = filledSlots.get(slot);
-		// Don't fill if a better timestamp was already filled
-		if (
-			filled &&
-			Math.abs(filled - slot) <= Math.abs(filled - frame.timestamp)
-		) {
-			continue;
-		}
-
-		drawSlot({
-			ctx,
-			frame,
-			filledSlots,
-			visualizationWidth,
-			timestamp: slot,
-			segmentDuration,
-			fromSeconds,
-		});
-	}
 };
 
 export const TimelineVideoInfo: React.FC<{
@@ -391,6 +150,7 @@ export const TimelineVideoInfo: React.FC<{
 				fromSeconds,
 				toSeconds,
 				aspectRatio: aspectRatio.current,
+				frameHeight: FILMSTRIP_HEIGHT,
 			});
 
 			fillWithCachedFrames({
@@ -400,6 +160,8 @@ export const TimelineVideoInfo: React.FC<{
 				src,
 				segmentDuration: toSeconds - fromSeconds,
 				fromSeconds,
+				devicePixelRatio: window.devicePixelRatio,
+				frameHeight: FILMSTRIP_HEIGHT,
 			});
 			repeatTarget();
 
@@ -430,6 +192,7 @@ export const TimelineVideoInfo: React.FC<{
 					toSeconds,
 					naturalWidth: targetWidth,
 					aspectRatio: aspectRatio.current,
+					frameHeight: FILMSTRIP_HEIGHT,
 				});
 
 				return Array.from(filledSlots.keys()).map(
@@ -468,6 +231,7 @@ export const TimelineVideoInfo: React.FC<{
 						toSeconds,
 						naturalWidth: targetWidth,
 						aspectRatio: aspectRatio.current,
+						frameHeight: FILMSTRIP_HEIGHT,
 					});
 					fillFrameWhereItFits({
 						ctx: targetCtx,
@@ -476,6 +240,8 @@ export const TimelineVideoInfo: React.FC<{
 						frame: transformed,
 						segmentDuration: toSeconds - fromSeconds,
 						fromSeconds,
+						devicePixelRatio: window.devicePixelRatio,
+						frameHeight: FILMSTRIP_HEIGHT,
 					});
 					repeatTarget();
 				} catch (e) {
@@ -502,6 +268,8 @@ export const TimelineVideoInfo: React.FC<{
 					src,
 					segmentDuration: toSeconds - fromSeconds,
 					fromSeconds,
+					devicePixelRatio: window.devicePixelRatio,
+					frameHeight: FILMSTRIP_HEIGHT,
 				});
 				repeatTarget();
 			})
