+++ packages/timeline-utils/README.md
@@ -1,0 +1,3 @@
+# @remotion/timeline-utils
+
+Undocumented utilities for timeline rendering.
