+++ packages/timeline-utils/eslint.config.mjs
@@ -1,0 +1,7 @@
+import {remotionFlatConfig} from '@remotion/eslint-config-internal';
+
+const config = remotionFlatConfig({react: false});
+
+export default {
+	...config,
+};
