+++ packages/timeline-utils/src/get-duration-or-compute.ts
@@ -1,0 +1,9 @@
+import type {Input} from 'mediabunny';
+
+export const getDurationOrCompute = async (input: Input) => {
+	return (
+		(await input.getDurationFromMetadata(undefined, {
+			skipLiveWait: true,
+		})) ?? input.computeDuration(undefined, {skipLiveWait: true})
+	);
+};
