+++ packages/timeline-utils/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../tsconfig.settings.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "dist"
+	},
+	"include": ["src"]
+}
